+++ app/src/main/java/edu/rpi/shuttletracker/core/ui/Errors.kt
@@ -1,4 +1,4 @@
-package edu.rpi.shuttletracker.ui.util
+package edu.rpi.shuttletracker.core.ui
 
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
