+++ app/src/main/java/edu/rpi/shuttletracker/feature/settings/SettingsViewModel.kt
@@ -1,11 +1,11 @@
-package edu.rpi.shuttletracker.ui.settings
+package edu.rpi.shuttletracker.feature.settings
 
 import androidx.compose.runtime.Immutable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import edu.rpi.shuttletracker.core.ui.theme.ThemeMode
 import edu.rpi.shuttletracker.data.repository.UserPreferencesRepository
-import edu.rpi.shuttletracker.ui.theme.ThemeMode
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
