+++ app/src/main/java/edu/rpi/shuttletracker/feature/setup/SetupPages.kt
@@ -1,4 +1,4 @@
-package edu.rpi.shuttletracker.ui.setup
+package edu.rpi.shuttletracker.feature.setup
 
 import androidx.compose.runtime.Composable
