+++ app/src/main/java/edu/rpi/shuttletracker/feature/setup/SetupScreenViewModel.kt
@@ -1,4 +1,4 @@
-package edu.rpi.shuttletracker.ui.setup
+package edu.rpi.shuttletracker.feature.setup
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
